# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/{expressions}": {
      "get": {
        "operationId": "calculate",
        "summary": "Calculate",
        "description": "Congestion Tax Calculation",
        "parameters": [
          {
            "name": "expressions",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request processed successfuly",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Calculator"
        ]
      }
    }
  },
  "info": {
    "title": "arithmetic-service",
    "description": "Arithmetic Service | [swagger.json](swagger.json)",
    "version": "0.0.1",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "BadRequestResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Bad request"
          },
          "status": {
            "type": "number",
            "example": 400
          }
        }
      },
      "InternalServerErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Internal Server Error"
          },
          "status": {
            "type": "number",
            "example": 500
          }
        }
      }
    }
  }
}